requirements.txt
streamlit>=1.20
pandas
wikipedia-api
google-generative-ai

# For testing
pytest
